Doc: Move generate doc get_parser to .cli.definition for each provider
- also unifiy the the provider CLI doc as cli-ref.rst
