+++ providers/fab/src/airflow/providers/fab/cli/definition.py
@@ -17,6 +17,7 @@
 from __future__ import annotations
 
 import textwrap
+from typing import TYPE_CHECKING
 
 from airflow.cli.cli_config import (
     ARG_DB_FROM_REVISION,
@@ -34,6 +35,9 @@ from airflow.cli.cli_config import (
     Arg,
     lazy_load_command,
 )
+
+if TYPE_CHECKING:
+    import argparse
 
 ############
 # # ARGS # #
@@ -365,3 +369,18 @@ def get_fab_cli_commands():
     ) >= packaging.version.parse("3.0.0"):
         commands.append(GroupCommand(name="fab-db", help="Manage FAB", subcommands=DB_COMMANDS))
     return commands
+
+
+def get_parser() -> argparse.ArgumentParser:
+    """
+    Generate documentation; used by Sphinx argparse.
+
+    :meta private:
+    """
+    from airflow.cli.cli_parser import AirflowHelpFormatter, DefaultHelpParser, _add_command
+
+    parser = DefaultHelpParser(prog="airflow", formatter_class=AirflowHelpFormatter)
+    subparsers = parser.add_subparsers(dest="subcommand", metavar="GROUP_OR_COMMAND")
+    for group_command in get_fab_cli_commands():
+        _add_command(subparsers, group_command)
+    return parser
