+++ providers/keycloak/docs/auth-manager/manage/permissions.rst
@@ -45,7 +45,7 @@ They also take the following optional parameters:
 
 * ``--dry-run``: If set, the command will check the connection to Keycloak and print the actions that would be performed, without actually executing them.
 
-Please check the `Keycloak auth manager CLI </cli-refs.html>`_ documentation for more information about accepted parameters.
+Please check the `Keycloak auth manager CLI </cli-ref.html>`_ documentation for more information about accepted parameters.
 
 One-go creation of permissions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
